better-reader-master/requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi 
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wincertstore==0.2
whitenoise==6.4.0
